Fixing Win32 x64 assembly implementation to save the full contents of the XMM6-15 registers.
It was using movq to store the XMM registers into the _mco_ctxbuf but this only stores the lowest 64 bits, leaving the top 64 bits as zero, so half of each XMM register was lost when switching contexts.

Fixing this by using movups, which stores the full 128 bits. If we could guarantee alignment, we could switch to movaps but this is safer at the moment and I don't see any solid guarantees that this is aligned at first glance.